Add video proccessing to the podcastShorter.py script
This commit adds the ability to process video files to the podcastShorter.py
You may drag and drop a video file onto the script and it will process it
and output a file with the same name but with the output.avi extension.

There is no audio processing in this commit, but it will be added in the future.