Split audio and video proccessing into two files
Now the audio and video processing is done in two separate files. This
will make it easier to maintain and understand the code.

The audio processing is done in audioProccesor.py and the video
processing is done in videoProcessor.py.

In this stage program will just output the same video as input.